refactor: separate alarm component in setting page
