+++ src/components/setting/index.ts
@@ -1,1 +1,3 @@
-export * from "./account";
+export * from "./account-setting";
+export * from "./alarm-setting";
+
